requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
arrow==1.4.0
asgiref==3.10.0
attrs==25.3.0
autobahn==25.10.2
Automat==25.4.16
av
binaryornot==0.4.4
certifi==2025.4.26
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.3.1
constantly==23.10.4
cookiecutter==2.6.0
cryptography==46.0.3
daphne==4.2.1
Django==5.2.3
django-browser-reload==1.21.0
django-channels==0.7.0
django-cors-headers==4.9.0
django-dotenv==1.4.2
django-htmx==1.23.0
django-phonenumber-field==8.1.0
django-tailwind==4.4.2
django-user-agents==0.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
gunicorn==23.0.0
hf-xet==1.1.3
honcho==2.0.0
huggingface-hub==0.33.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgpack==1.1.2
multidict==6.4.4
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
phonenumbers==9.0.7
pillow==11.2.1
propcache==0.3.2
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytailwindcss==0.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
redis==7.0.1
regex==2024.11.6
reportlab==4.4.1
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.2.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
stripe==14.1.0
style==1.1.0
text-unidecode==1.3
tokenizers==0.21.1
tqdm==4.67.1
twilio==9.6.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.14.0
tzdata==2025.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
update==0.0.1
urllib3==2.4.0
user-agents==2.2.0
vobject==0.9.9
whitenoise==6.9.0
yarl==1.20.1
zope.interface==8.1
